+++ CHANGELOG.md
@@ -8,8 +8,11 @@ upgrades, e.g. `v0.6.x` -> `v0.7.x`. The config file format and
 [API](ref/api.md) currently have no stability guarantees, so they may change
 even on minor releases, e.g. `v0.7.5` -> `v0.7.6`.
 
-## unreleased
+## v0.7.9 (2023-10-21)
 
+*  upgrade date/time pickers to `@mui/x-date-pickers` v6 beta, improving
+   time entry. Fixes
+   [#256](https://github.com/scottlamb/moonfire-nvr/issues/256).
 *  `systemd` integration on Linux
    *   notify `systemd` on starting/stopping. To take advantage of this, you'll
        need to modify your `/etc/systemd/moonfire-nvr.service`. See